chore: add tests for classifyDestinations function
🔒 Scanned for secrets using gitleaks 8.28.0
